// IQFront/src/app/modelos/listaClientes.interface.ts
export interface ListaClientesI{
      
    id_cliente: string;
    razon_social: string;
    identificacion: string;
    nombre_comercial: string;
    direccion: string;
    telefono: string;
    telefono2: string;
    codigo_postal: string;
    tipo_cliente: string;
    nivel: string;
    estado_cliente: string;
    ciudad_id: string;
    
}